% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concoR.R
\name{block}
\alias{block}
\alias{block_names}
\alias{concor}
\title{Hierarchical clustering using CONCOR.}
\usage{
block(m0, cutoff = 0.999, max.iter = 25, p = 1)

concor(m0, cutoff = 0.999, max.iter = 50)

block_names(p_list)
}
\arguments{
\item{m0}{A list of \eqn{n \times n} matrices, each of which refers to a
relation on a set of vertices.  Only one-mode data are supported at present.
#'}

\item{cutoff}{A value between 0 and 1 used to determine convergence.}

\item{max.iter}{An integer representing the maximum number of iterations.}

\item{p}{An integer representing the desired number of partitions.}
}
\value{
A \code{data.frame} depicting the block assignment for each vertex.
}
\description{
Paritions relational data using a CONvergence of iterated CORrelations
(CONCOR) algorithm.
}
\examples{
data(bank_wiring)
b <- block(bank_wiring, p = 2)
lapply(bank_wiring, block_dens, blocks = b)
}
\references{
Breiger, R.L., Boorman, S.A., and Arabie, P.  1975.  An Algorithm
for Clustering Relational Data with Applications to Social Network Analysis
and Comparison with Multidimensional Scaling.  \emph{Journal of Mathematical
Psychology}, 12: 328--383.
}

