% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/block_dens.R
\name{block_dens}
\alias{block_dens}
\title{Calculate tie densities.}
\usage{
block_dens(m0, blocks, digits = 3)
}
\arguments{
\item{m0}{A list of \eqn{n \times n} matrices, each of which refers to a
relation on a set of vertices.  Only one-mode data are supported at present.}

\item{blocks}{A \code{data.frame} depicting the assignment of vertices
to blocks.}

\item{digits}{The number of significant digits used when printing results.}
}
\value{
A \code{list} of matrices representing block-by-block tie densities
for each relation containined in \code{m0}.
}
\description{
Uses tie density to assess the fit of a blockmodel on multiple relations.
}
\examples{
data(bank_wiring)
b <- block(bank_wiring, p = 2)
lapply(bank_wiring, block_dens, blocks = b)
}
\references{
Breiger, R.L., Boorman, S.A., and Arabie, P.  1975.  An Algorithm
for Clustering Relational Data with Applications to Social Network Analysis
and Comparison with Multidimensional Scaling.  \emph{Journal of Mathematical
Psychology}, 12: 328--383.
}

