% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concoR.R
\name{concor_hca}
\alias{block_names}
\alias{concor}
\alias{concor_hca}
\title{Hierarchical clustering using CONCOR.}
\usage{
concor_hca(m0, cutoff = 0.999, max.iter = 25, p = 1)

concor(m0, cutoff = 0.999, max.iter = 50)

block_names(p_list)
}
\arguments{
\item{m0}{A list of \eqn{n} x \eqn{n} matrices, each of which refers to a
relation on a set of vertices.  Only one-mode data are supported at present.}

\item{cutoff}{A value between 0 and 1 used to determine convergence.}

\item{max.iter}{An integer representing the maximum number of iterations.}

\item{p}{An integer representing the desired number of partitions.}
}
\value{
A \code{data.frame} depicting the block assignment for each vertex.
Rows are sorted to match the order in which vertices appear in the original
data.  The column vector indicating block membership can be passed directly to
the \code{blockmodel} command in \code{sna}.
}
\description{
Paritions relational data using a CONvergence of iterated CORrelations
(CONCOR) algorithm.
}
\examples{
data(bank_wiring)
library(sna)
b <- concor_hca(bank_wiring, p = 2)
g <- as.sociomatrix.sna(bank_wiring)
bm <- blockmodel(g, ec = b$block, glabels = names(bank_wiring))
plot(bm)
}
\references{
Breiger, R.L., Boorman, S.A., and Arabie, P.  1975.  An Algorithm
for Clustering Relational Data with Applications to Social Network Analysis
and Comparison with Multidimensional Scaling.  \emph{Journal of Mathematical
Psychology}, 12: 328--383.
}

